+++ src/main/java/net/frozenblock/wilderwild/worldgen/modification/WWWorldGen.java
@@ -378,7 +378,7 @@ public final class WWWorldGen {
 				BiomeSelectors.all(),
 				(biomeSelectionContext, context) -> {
 					BiomeModificationContext.GenerationSettingsContext generationSettings = context.getGenerationSettings();
-					if (biomeSelectionContext.hasTag(WWBiomeTags.HAS_POLLEN)) {
+					if (WWWorldgenConfig.GENERATE_POLLEN && biomeSelectionContext.hasTag(WWBiomeTags.HAS_POLLEN)) {
 						generationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, WWPlacedFeatures.POLLEN_PLACED.getKey());
 					}
 				});